+++ test/__fixtures/utils/test_helper.js
@@ -15,6 +15,8 @@ function TestHelper() {
   this.__happnerInstances = {};
   this.TCPProxy = require('./tcp-proxy/proxy');
   this.expect = require('expect.js');
+  this.package = require('../../../package.json');
+  this.happnPackage = JSON.parse(fs.readFileSync(path.resolve(__dirname, '../../../node_modules/happn-3/package.json')));
 }
 
 TestHelper.create = function(){